feat: 공통 헤더 적용, 디렉터리 정리
