+++ src/pages/address/Address.module.css
@@ -4,8 +4,6 @@
   width: 100%;
   margin: 0 auto;
   font-size: 32px;
-  font-weight: bold;
-  font-family: sans-serif;
 }
 .header{
   text-align: center;
