+++ src/pages/address/AddressEdit.module.css
@@ -1,5 +1,6 @@
 .container {
   padding: 20px;
+  padding-top: 84px;
   font-family: sans-serif;
 }
 
@@ -38,7 +39,7 @@
   padding: 6px 12px;
   border-radius: 16px;
   font-size: 14px;
-  border: 1px solid #ccc;
+  border: 1px solid var(--border-color);
 }
 .iconWithContent {
   display: flex;
@@ -53,11 +54,12 @@
 }
 
 .addressTextGroup {
-  margin-bottom: 16px;
+  margin-bottom: 12px;
 }
 
 .primaryAddress {
-  font-weight: bold;
+  margin-bottom: 8px;
+  font-weight: 600;
   font-size: 16px;
 }
 
@@ -68,46 +70,35 @@
 }
 
 .wowArea {
-  margin-top: 6px;
+  margin-top: 12px;
   font-size: 12px;
 }
 
 .wow {
-  background-color: #2B5C9F;
+  background-color: var(--wow-color);
   color: white;
-  padding: 1.5px 4px 0.5px;
+  padding: 1px 4px 1px;
   border-radius: 12px;
   font-size: 10px;
-  font-weight: bold;
   margin-right: 4px;
   margin-left: 4px;
 }
 
 .wowText {
-  color: #2B5C9F;
+  color: var(--wow-color);
   font-weight: bold;
 }
 
 .detailInput {
-  width: 96%;
-  height: 40px;
-  padding: 10px;
+  width: 100%;
+  height: 60px;
+  padding: 16px;
+  margin-bottom: 12px;
   font-size: 16px;
   border: 1px solid #ccc;
   border-left: none;
   border-right: none;
-}
-
-.detailInput2 {
-  width: 96%;
-  height: 40px;
-  padding: 10px;
-  font-size: 16px;
-  border: 1px solid #ccc;
-  border-top: none;
-  border-left: none;
-  border-right: none;
-  margin-bottom: 20px;
+  box-sizing: border-box;
 }
 
 .labelButtonGroup {
@@ -132,7 +123,7 @@
 
 .labelButton.selected {
   background-color: #e5f4ff;
-  border-color: #2196f3;
+  border-color: var(--theme-color);
 }
 
 .labelIcon {
@@ -144,7 +135,7 @@
 .submitButton {
   width: 100%;
   padding: 14px;
-  background-color: #00aaff;
+  background-color: var(--theme-color);
   color: white;
   border: none;
   border-radius: 10px;
